modules/cloudFunc/requirements_dump.txt
# requirements.txt
functions-framework==3.*
google-api-python-client==2.*
google-auth==2.*